Remark how the lazy-apps setting fixes a startup issue.
With uWSGI+nginx, the first request that I do (per worker) did not
generate a response. The slackbridge daemon doesn't mind, but the
calling process only gets an empty response after a minute.

For the outgoing slack hook, that is fine, because it was a 200; it
won't try to retransmit. But it's confusing nevertheless.

For some reason, the forking/pipe setup in the first application call
breaks communication with nginx.

Tested versions: uwsgi 1.9.17.1-5build5, nginx 1.4.6-1ubuntu3.2.

Current uwsgi config:

    [uwsgi]
    wsgi-file = /srv/http/my.example.com/slackbridge.py
    # fix so relative imports work
    chdir = /srv/http/my.example.com
    lazy-apps = true
    workers = 2