Refactor slackbridge.py to wsgi.py. Add config loader classes.
If you use the PORTAL_n_SIDE_n env config or the slackbridgeconf.py
style config, you need only replace your reference to slackbridge.py to
wsgi.py.

However, you may now start using an inifile for configuration. See
sample.ini for an example.

Configuration loading is done in the following order:

- environment (PORTAL_n_SIDE_n_...)
- if that was not found, load inifile (./slackbridge.ini or the
  filename found in SLACKBRIDGE_INIFILE env)

If slackbridgeconf.py is found, it is loaded as well. Expect this to
change in future commits.

If you're replacing your CONFIG-style slackbridge conf with either the
INI or ENV style config, you can use testconfig.py to compare the
output. The old CONFIG-style config was write only, so a double-check to
see if you made any mistakes in the new INI or ENV config is highly
recommended.

Touches issue #12.