errors: Quick and dirty timers
In our case, the ERROR mail send failed, causing a gigantic queue to
build up. This subprocess alarm won't report failed mail, but will
resume with the next message. And report all other timeouts by mail.

The ipc.send() should also be helped by an alarm, but when I tested it
with a time.sleep in the try-block, the alarm did not fire at all. No
idea why that is yet. (Are threads enabled anyway and is this process
not the main thread?)