attachments: Quick hack to report that we'r NOT handling non-text
If you send an image/attachment without text, you'll now get a message
that it cannot get sent.

The outgoing webhook only gives us text. And only if there is none, do
we realise that someone tried to send non-text.

For attachments *with* text, this heuristic fails and no one will be
notified that the image/attachment does not reach the peer side. (This
is unchanged/untouched by this changeset.)

See also #14